Small cleanup and improvements - move hooks to form folder - replace `type + &` with `interface + extends` - improve package.json and more
