+++ packages/ui-informed/src/form/Form.tsx
@@ -1,32 +1,31 @@
-import { cn } from '@utima/ui';
 import { Form as InformedForm } from 'informed';
 import { useMemo, type ComponentProps } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { ZodObject, ZodRawShape } from 'zod';
 
-import { FormContext, type FormContextType } from '../hooks/useFormContext';
 import {
-  useFormSubmitActions,
+  useDefaultSubmitActions,
   type TypedFormState,
-} from '../hooks/useFormSubmitActions';
+} from './useDefaultSubmitActions';
+import { FormContext, type FormContextType } from './useFormContext';
 
-export type FormProps<T> = Omit<
-  ComponentProps<typeof InformedForm>,
-  'initialValues' | 'onSubmit'
-> & {
+export interface FormProps<T>
+  extends Omit<
+    ComponentProps<typeof InformedForm>,
+    'initialValues' | 'onSubmit'
+  > {
   initialValues: T;
   disabled?: boolean;
   readOnly?: boolean;
   loading?: boolean;
-  columns?: boolean;
   onSubmit?: (formState: TypedFormState<T>) => Promise<unknown>;
   zodSchema?: ZodObject<ZodRawShape>;
-};
+}
 
 /**
- * Basic form component used to wrap form inputs. It provides action bar,
- * column layout, and wraps all content in Informed Form component,
- * which handles form state and validation.
+ * Basic form component used to wrap form inputs. It wraps
+ * all content in Informed Form component, which handles
+ * form state and validation.
  */
 export function Form<T>({
   onSubmit,
@@ -35,13 +34,12 @@ export function Form<T>({
   readOnly = false,
   disabled = false,
   loading = false,
-  columns = true,
   zodSchema,
   ...restProps
 }: FormProps<T>) {
   const { t } = useTranslation();
   const { handleSubmit, handleSubmitFailure, isSubmitting } =
-    useFormSubmitActions<T>({
+    useDefaultSubmitActions<T>({
       onSubmit,
     });
 
@@ -67,16 +65,7 @@ export function Form<T>({
       {...restProps}
     >
       <FormContext.Provider value={contextValue}>
-        <div
-          className={cn(columns && 'sm:grid gap-6')}
-          style={{
-            gridTemplateColumns: columns
-              ? 'repeat(11, minmax(0, 1fr))'
-              : undefined,
-          }}
-        >
-          {children}
-        </div>
+        {children}
       </FormContext.Provider>
     </InformedForm>
   );
